chore: split DatePicker spec into 2 parts
