+++ package.json
@@ -56,7 +56,6 @@
     "test:main:suite-2": "yarn workspace @ui5/webcomponents test:suite-2",
     "test:main:suite-3": "yarn workspace @ui5/webcomponents test:suite-3",
     "test:main:suite-input": "yarn workspace @ui5/webcomponents test:suite-input",
-    "test:main:suite-datetime": "yarn workspace @ui5/webcomponents test:suite-datetime",
     "test:fiori": "yarn workspace @ui5/webcomponents-fiori test",
     "test": "npm-run-all --sequential test:base test:main test:fiori",
     "lint": "wsrun --exclude-missing lint",
