+++ packages/main/test/specs/input/Calendar.spec.js
