+++ packages/main/test/specs/input/DateRangePicker.spec.js
