+++ packages/main/test/specs/input/DayPicker.spec.js
